Frontend: Split forms module
