+++ orthos2/frontend/forms/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module groups together all Django Forms that the application needs. Forms for the API should be living inside the
+"orthos2.api.forms" package.
+"""
